Actually fix gen_ai nested fields patching

+++ ecs/generator/images/generator.sh
@@ -67,13 +67,20 @@ fix_gen_ai_nested_fields() {
   local in_file="$1"
   local csv_file="$2"
 
-  echo "Fixing gen_ai nested fields in $in_file"
-  jq '
-    (.mappings.properties.gen_ai.properties.request.properties.encoding_formats?.type) = "keyword" |
-    (.mappings.properties.gen_ai.properties.request.properties.stop_sequences?.type) = "keyword" |
-    (.mappings.properties.gen_ai.properties.response.properties.finish_reasons?.type) = "keyword"
-    ' "$in_file" >"${in_file}.tmp"
+  # Only apply changes if any of the targeted gen_ai fields exist in the template
+  if jq -e 'has("mappings") and (
+      .mappings.properties.gen_ai?.properties.request?.properties.encoding_formats? != null or
+      .mappings.properties.gen_ai?.properties.request?.properties.stop_sequences? != null or
+      .mappings.properties.gen_ai?.properties.response?.properties.finish_reasons? != null
+    )' "$in_file" >/dev/null 2>&1; then
+    echo "Fixing gen_ai nested fields in $in_file"
+    jq '
+      (.mappings.properties.gen_ai.properties.request.properties.encoding_formats?.type) = "keyword" |
+      (.mappings.properties.gen_ai.properties.request.properties.stop_sequences?.type) = "keyword" |
+      (.mappings.properties.gen_ai.properties.response.properties.finish_reasons?.type) = "keyword"
+      ' "$in_file" >"${in_file}.tmp"
     mv "${in_file}.tmp" "$in_file"
+  fi
 
   echo "Fixing gen_ai nested fields in $csv_file"
   sed -i 's/encoding_formats,nested,extended,,/encoding_formats,keyword,extended,array,/g' "$csv_file"
@@ -129,11 +136,7 @@ generate_mappings() {
   echo "Deleting \"synthetic_source_keep\" mappings setting from the index template"
   sed -i '/synthetic_source_keep/d' "$in_file"
 
-
-  # Only apply fix_gen_ai_nested_fields to stateless/* modules (not to stateful/* or others)
-  if [[ "$ecs_module" == stateless/* ]]; then
-    fix_gen_ai_nested_fields "$in_file" "$csv_file"
-  fi
+  fix_gen_ai_nested_fields "$in_file" "$csv_file"
 
   if [[ "$ecs_module" != stateless* ]]; then
     # Delete the "tags" field from the index template
